+++ code/cell.modules/pkg.sys/ui.primitives/src/components/EventStack/factory.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import { EventStackCard } from '../EventStack.Card';
+import { EventStackCardFactory } from './types';
+
+export const cardFactory: EventStackCardFactory = (args) => {
+  return (
+    <EventStackCard
+      key={args.id}
+      count={args.count}
+      event={args.event}
+      width={args.width}
+      isTopCard={args.isTopCard}
+      showPayload={args.showPayload}
+      onShowPayloadToggle={args.toggleShowPayload}
+    />
+  );
+};
