+++ code/cell.modules/pkg.sys/ui.primitives/src/components/EventStack/index.ts
@@ -1,1 +1,2 @@
 export * from './EventStack';
+export * from './types';
