+++ code/cell.modules/pkg.sys/ui.primitives/src/components/EventStack/types.ts
@@ -1,0 +1,11 @@
+import { t } from '../../common';
+
+export type EventStackEvent = { id: string; event: t.Event; count: number };
+
+export type EventStackCardFactory = (args: EventStackCardFactoryArgs) => JSX.Element;
+export type EventStackCardFactoryArgs = EventStackEvent & {
+  width: number;
+  isTopCard: boolean;
+  showPayload?: boolean;
+  toggleShowPayload?: () => void;
+};
